Resolved Issue #41
Added ability to set the webrtc config (for when enable is called) in
the default setRtcommEndpointConfig or when enable is called.  This now includes
a 'iceServers' which is is an ARRAY of strings that would match an
iceserver like:  "stun:stun.l.google.com:19302.

This changes the setRtcommEndpointConfig API.  It now supports ALL
options that can be set on the WebRTCConnection in enable() via a
'webrtcConfig' tag.